chore: Typescript 프로젝트로 초기 세팅 완료(husky만 적용)
