+++ src/App.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function App() {
+  return <div>App.tsx</div>;
+}
+
+export default App;
