--- src/containers/CommentListContainer.js
@@ -1,8 +1,0 @@
-import React from "react";
-import CommentList from "../components/CommentList";
-
-function CommentListContainer() {
-  return <CommentList />;
-}
-
-export default CommentListContainer;
