--- src/containers/FormContainer.js
@@ -1,8 +1,0 @@
-import React from "react";
-import Form from "../components/Form";
-
-function FormContainer() {
-  return <Form />;
-}
-
-export default FormContainer;
