--- src/containers/PageListContainer.js
@@ -1,8 +1,0 @@
-import React from "react";
-import PageList from "../components/PageList";
-
-function PageListContainer() {
-  return <PageList />;
-}
-
-export default PageListContainer;
