--- src/util/async.utill.js
@@ -1,22 +1,0 @@
-export const reducerUtils = {
-  initial: (data = null) => ({
-    loading: false,
-    data,
-    error: null,
-  }),
-  loading: (preveState = null) => ({
-    data: preveState,
-    loading: true,
-    error: null,
-  }),
-  success: (data) => ({
-    data,
-    loading: false,
-    error: null,
-  }),
-  error: (error) => ({
-    data: null,
-    loading: false,
-    error,
-  }),
-};
